Export iliad chain
