+++ packages/core-sdk/src/utils/utils.ts
@@ -10,10 +10,10 @@ import {
   isAddress,
   checksumAddress,
   Address,
-  defineChain,
 } from "viem";
 
 import { SupportedChainIds } from "../types/config";
+import { iliad } from "./chain";
 
 export async function waitTxAndFilterLog<
   const TAbi extends Abi | readonly unknown[],
@@ -76,31 +76,7 @@ export async function waitTx(
     ...params,
   });
 }
-//TODO: Some information is waiting for confirmation about chain
-export const iliad = defineChain({
-  id: 15_13,
-  name: "iliad",
-  nativeCurrency: { name: "IP", symbol: "IP", decimals: 18 },
-  rpcUrls: {
-    default: {
-      http: ["https://testnet.storyrpc.io"],
-      webSocket: ["wss://story-network.rpc.caldera.xyz/ws"],
-    },
-  },
-  blockExplorers: {
-    default: {
-      name: "Explorer",
-      url: "https://testnet.storyscan.xyz",
-    },
-  },
-  contracts: {
-    multicall3: {
-      address: "0xcA11bde05977b3631167028862bE2a173976CA11",
-      blockCreated: 5882,
-    },
-  },
-  testnet: true,
-});
+
 export function chainStringToViemChain(chainId: SupportedChainIds): Chain {
   switch (chainId) {
     case "1513":
